Move save changes handler into handlers file
