+++ src/local/util/dom/events/handlers.js
@@ -1,13 +1,15 @@
 import { GLOBALS } from "../../../../globals";
-import { insertFavicon } from "..";
+import { insertFavicon, getBodyBackgroundColor } from "..";
 import { onUpdateBodyAlign } from "../align";
-import { isImageFile } from "../../files";
+import { isImageFile, saveFile } from "../../files";
+import { enableLocalControls } from "../../../enable";
 import {
   CANCEL_FAVICON_UPDATE_ID,
   CONFIRM_FAVICON_UPDATE_ID,
   CURRENT_SOCIAL_IMAGE_ID,
   CURRENT_SOCIAL_IMAGE_PREVIEW_ID,
   CURRENT_TWITTER_IMAGE_ID,
+  LOCAL_CONTROLS_ID,
   STRINGS,
   UPDATE_BODY_ALIGN_CONTAINER,
   UPDATE_BODY_ALIGN_ID,
@@ -15,6 +17,37 @@ import {
   WIDTH_SLIDER_ID,
   WIDTH_SLIDER_VALUE_ID,
 } from "../../../constants";
+
+export async function onClickSaveChanges(_event) {
+  const localControls = document.getElementById(LOCAL_CONTROLS_ID);
+  localControls.remove();
+  const htmlSourceCode = `
+            <!DOCTYPE html>
+            <html lang="${document.documentElement.lang}"  id="html-element">
+              <head>
+                ${document.head.innerHTML}
+              </head>
+              <body
+                style="
+                  background-color: ${getBodyBackgroundColor()};
+                  text-align: ${document.body.style.textAlign};
+                  align-items: ${document.body.style.alignItems};
+                ">
+                ${document.body.innerHTML}
+              </body>
+            </html>
+          `;
+  try {
+    await saveFile(htmlSourceCode);
+  } catch (error) {
+    if (error?.message === "The user aborted a request.") {
+      // Error is just cancel button
+    } else {
+      alert(error?.message ?? "Error saving file.");
+    }
+  }
+  enableLocalControls();
+}
 
 export function onUpdateImgElementAlign(value) {
   if (value === "default") {
